+++ src/include/ecrs_core.h
@@ -162,7 +162,7 @@ typedef struct {
  *  match the query
  */
 int fileBlockEncode(const DBlock * data,
-		    unsigned int len,
+		    unsigned short len,
 		    const HashCode512 * query,
 		    Datastore_Value ** value);
 
@@ -171,7 +171,7 @@ int fileBlockEncode(const DBlock * data,
  * a certain block of data.
  */
 void fileBlockGetQuery(const DBlock * data,
-		       unsigned int len,
+		       unsigned short len,
 		       HashCode512 * query);
 
 
@@ -180,13 +180,13 @@ void fileBlockGetQuery(const DBlock * data,
  * a certain block of data.
  */
 void fileBlockGetKey(const DBlock * data,
-		     unsigned int len,
+		     unsigned short len,
 		     HashCode512 * key);
 
 /**
  * What is the type of the given block of data?
  */
-unsigned int getTypeOfBlock(unsigned int size,
+unsigned int getTypeOfBlock(unsigned short size,
 			    const DBlock * data);
 
 /**
@@ -200,7 +200,7 @@ unsigned int getTypeOfBlock(unsigned int size,
  * @return SYSERR if the content is invalid or
  *   the content type is not known
  */
-int getQueryFor(unsigned int size,
+int getQueryFor(unsigned short size,
 		const DBlock * data,
 		HashCode512 * query);
 
@@ -218,7 +218,7 @@ int getQueryFor(unsigned int size,
  *         query type
  */
 int isDatumApplicable(unsigned int type,
-		      unsigned int size,
+		      unsigned short size,
 		      const DBlock * data,
 		      unsigned int keyCount,
 		      const HashCode512 * keys);
